Split package lists across config files
This way, instead of having one central place where
installed user/system programs are installed, we have
different modules in which the packages being configured
are installed alongside the config as well.

This also means certain configs are obsolete or consolidated.

Now there are some ground rules for adding packages:

- Shell utils: home/shell.nix
- Generic user programs: home/default.nix
- System utils (they only run as root): hosts/kratos/systemprograms.nix
- System utils with configs: system/<util>.nix (or wherever the config is)
- User util with config: home/<util>.nix (or wherever the config is)